+++ src/main/kotlin/com/scrumdapp/checkpointservice/dto/CheckpointDto.kt
@@ -1,14 +1,14 @@
 package com.scrumdapp.checkpointservice.dto
 
 import com.fasterxml.jackson.annotation.JsonInclude
-import com.scrumdapp.checkpointservice.entities.Checkpoint
+
 
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 data class CheckpointResponseDto(
     val id: Long,
-    val sessionId: Int,
-    val groupUserId: Int,
+    val sessionId: Long,
+    val groupUserId: Long,
     val presence: Int?,
     val impediment: String?,
     val stars: Int?,
